EfficientGEBD/requirements.txt
einops==0.8.0
mmaction2==1.2.0
mmcv==2.1.0
mmengine==0.9.1
numpy==1.26.4
opencv_contrib_python==4.5.4.60
opencv_python==4.5.4.60
opencv_python_headless==4.5.4.60
pickle5==0.0.11
Pillow==10.1.0
protobuf==3.20.0
ptflops==0.7.2.2
scikit-learn==1.2.0
tabulate==0.9.0
tensorboard==2.9.0
timm==0.4.12
torch==2.1.2
torchvision==0.16.2
tqdm==4.65.0
transformers==4.39.1
yacs==0.1.8
